requirements.txt
gpt4all==0.2.3
chromadb==0.3.23
urllib3==2.0.2
pdfminer.six==20221105
python-dotenv==1.0.0
unstructured==0.6.6
extract-msg==0.41.1
tabulate==0.9.0
pandoc==2.3
pypandoc==1.11
pytesseract
pypdfium2
tqdm==4.65.0
twine
build

datasets
bitsandbytes
langchain
peft
text-generation
torch

runpod==0.10.0
requests==2.31.0 
text-generation==0.6.0 
huggingface_hub==0.16.4
nvidia-smi